Async Await implementation

+++ server/server.js
@@ -61,16 +61,17 @@ app.get('/todos/:id',authenticate ,(req, res) => {
     ).catch((e) => res.status(400).send());
 });
 
-app.delete('/todos/:id',authenticate,(req,res) => {
-    let id = req.params.id;
+app.delete('/todos/:id',authenticate, async (req,res) => {
+    const id = req.params.id;
     if(!ObjectID.isValid(id)) return res.status(404).send();
-    Todo.findOneAndRemove({
-        _id: id,
-        _creator: req.user._id
-    }).then((todo) => {
+
+    try {
+        const todo = await Todo.findOneAndRemove({_id: id,_creator: req.user._id});
         if(!todo) return res.status(404).send();
         res.send({todo});
-    }).catch(e => res.status(400).send());;
+    } catch (e) {
+        res.status(400).send()
+    }
 });
 
 app.patch('/todos/:id', authenticate, (req,res) => {
@@ -94,39 +95,40 @@ app.patch('/todos/:id', authenticate, (req,res) => {
 
 });
 
-app.post('/users', (req, res) => {
-    let body = _.pick(req.body,['email','password']);
-    let user = new User(body);
-
-    user.save().then(
-        (user) => { 
-            return user.generateAuthToken();
-        }
-    ).then(
-        (token) => {
-            res.header('x-auth', token).send(user);
-        }
-    ).catch((error) => res.status(400).send({error}));
+app.post('/users', async (req, res) => {
+    try {
+        const body = _.pick(req.body,['email','password']);
+        const user = new User(body);
+        await user.save()
+        const token = await user.generateAuthToken();
+        res.header('x-auth', token).send(user);
+    } catch (e) {
+        res.status(400).send({e});
+    }
 });
 
 app.get('/users/me',authenticate,(req, res) => {
     res.send(req.user);
 });
 
-app.post('/users/login', (req,res) => {
-    let body = _.pick(req.body,['email', 'password']);
-
-    User.findByCredentials(body.email, body.password).then((user) => {
-        return user.generateAuthToken().then((token) => {
-            res.header('x-auth', token).send(user);
-        });
-    }).catch((e) => res.status(400).send({}))
+app.post('/users/login', async (req,res) => {
+    try {
+        const body = _.pick(req.body,['email', 'password']);
+        const user = await User.findByCredentials(body.email, body.password);
+        const token = await user.generateAuthToken();
+    res.header('x-auth',token).send(user);
+    } catch (e) {
+        res.status(400).send({})
+    }
 });
 
-app.delete('/users/me/token', authenticate,(req,res) => {
-    req.user.removeToken(req.token).then(
-    (success) => res.status(200).send(),
-    (error) => res.status(400).send())
+app.delete('/users/me/token', authenticate, async (req,res) => {
+    try {
+        await req.user.removeToken(req.token);
+        res.status(200).send();
+    } catch (e) {
+        res.status(400).send()
+    }
 });
 
 app.listen(port, () => {
